Moved git files into the main bigbang library. Also renamed the files to be lowercase with underscores. Also, the git repos will now go into archives/sample_git_repos
