--- git_data/git_locals.json
@@ -1,5 +1,0 @@
-{
-    "npm": "sample_git_repos/npm", 
-    "scipy": "sample_git_repos/scipy", 
-    "django": "sample_git_repos/django"
-}
